requirements.txt
fastmcp==1.0
mcp-neo4j-cypher
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.1
neo4j>=4.4.0
pandas>=1.3.3
numpy>=1.21.2
matplotlib>=3.4.3
networkx>=2.6.3
